Improve processing of author names
In the MEDLINE AU field, the first initials are always a single string, so there should never be a multi-word first name. If there is more than one string preceding the first initials and lineage, it means that the last name contains multiple words. This edit allows for correct processing of multi-word last names. If one of the words in the last name is a single character, it is given a '.' because it is probably an abbreviation.

See for example the multi-word last names in these MEDLINE entries:
  http://www.ncbi.nlm.nih.gov/pubmed?Db=pubmed&Cmd=Retrieve&list_uids=24701472&dopt=medline
  http://www.ncbi.nlm.nih.gov/pubmed?Db=pubmed&Cmd=Retrieve&list_uids=24701509&dopt=medline